backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
httpx
pydantic
pandas
numpy
scikit-learn
tensorflow
yfinance
matplotlib
